Move util/lightbox-router to ./lightbox-router
